app/requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
psycopg2-binary==2.9.7
redis==4.6.0
Werkzeug==2.3.7
Flask-WTF==1.1.1
email-validator==2.0.0
flask-bcrypt==1.0.1
msal==1.24.1
requests==2.31.0